chore: remove all unused methods and move non-public classes into core package.
