+++ core/src/main/java/com/google/cloud/sql/core/CoreSocketFactory.java
@@ -19,7 +19,6 @@ package com.google.cloud.sql.core;
 import com.google.api.client.http.HttpRequestInitializer;
 import com.google.cloud.sql.AuthType;
 import com.google.cloud.sql.CredentialFactory;
-import com.google.cloud.sql.SqlAdminApiFetcherFactory;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.ListenableFuture;
@@ -55,14 +54,6 @@ import jnr.unixsocket.UnixSocketChannel;
 public final class CoreSocketFactory {
 
   public static final String CLOUD_SQL_INSTANCE_PROPERTY = "cloudSqlInstance";
-  /**
-   * Property used to set the application name for the underlying SQLAdmin client.
-   *
-   * @deprecated Use {@link #setApplicationName(String)} to set the application name
-   *     programmatically.
-   */
-  @Deprecated public static final String USER_TOKEN_PROPERTY_NAME = "_CLOUD_SQL_USER_TOKEN";
-
   public static final String DEFAULT_IP_TYPES = "PUBLIC,PRIVATE";
   private static final String UNIX_SOCKET_PROPERTY = "unixSocketPath";
   private static final Logger logger = Logger.getLogger(CoreSocketFactory.class.getName());
@@ -114,10 +105,6 @@ public final class CoreSocketFactory {
               executor);
     }
     return coreSocketFactory;
-  }
-
-  static int getDefaultServerProxyPort() {
-    return DEFAULT_SERVER_PROXY_PORT;
   }
 
   // TODO(kvg): Figure out better executor to use for testing
@@ -210,21 +197,6 @@ public final class CoreSocketFactory {
     return getInstance().getCloudSqlInstance(csqlInstanceName, AuthType.PASSWORD).getSslData();
   }
 
-  /** Returns data that can be used to establish Cloud SQL SSL connection. */
-  static SslData getSslData(String csqlInstanceName, AuthType authType) {
-    return getInstance().getCloudSqlInstance(csqlInstanceName, authType).getSslData();
-  }
-
-  /** Returns data that can be used to establish Cloud SQL SSL connection. */
-  public static SslData getSslData(String csqlInstanceName) throws IOException {
-    return getSslData(csqlInstanceName, false);
-  }
-
-  /** Returns default ip address that can be used to establish Cloud SQL connection. */
-  public static String getHostIp(String csqlInstanceName) {
-    return getInstance().getHostIp(csqlInstanceName, listIpTypes(DEFAULT_IP_TYPES));
-  }
-
   /** Returns preferred ip address that can be used to establish Cloud SQL connection. */
   public static String getHostIp(String csqlInstanceName, String ipTypes) throws IOException {
     return getInstance().getHostIp(csqlInstanceName, listIpTypes(ipTypes));
@@ -291,24 +263,11 @@ public final class CoreSocketFactory {
   }
 
   /** Returns the current User-Agent header set for the underlying SQLAdmin API client. */
-  public static String getApplicationName() {
+  private static String getApplicationName() {
     if (coreSocketFactory != null) {
       return coreSocketFactory.adminApiService.getApplicationName();
     }
-    return System.getProperty(USER_TOKEN_PROPERTY_NAME, "");
-  }
-
-  /**
-   * Sets the User-Agent header for requests made using the underlying SQLAdmin API client.
-   *
-   * @throws IllegalStateException if the SQLAdmin client has already been initialized
-   */
-  public static void setApplicationName(String applicationName) {
-    if (coreSocketFactory != null) {
-      throw new IllegalStateException(
-          "Unable to set ApplicationName - SQLAdmin client already initialized.");
-    }
-    System.setProperty(USER_TOKEN_PROPERTY_NAME, applicationName);
+    return "";
   }
 
   /**
